Completed all class structures. TODO blocks added(to be expanded) Implemented User and cash-related classes. Added ID to Account.

+++ phase1/src/atm/AssetAccount.java
@@ -6,7 +6,7 @@ abstract class AssetAccount extends Account implements DepositableAccount, Withd
     }
 
     @Override
-    int getNetBalance() {
+    double getNetBalance() {
         return getBalance();
     }
 }
