+++ phase1/src/atm/LineOfCreditAccount.java
@@ -6,7 +6,7 @@ class LineOfCreditAccount extends DebtAccount implements WithdrawableAccount {
 
     @Override
         /* TODO Implementation */
-    int getBalance() {
+    double getBalance() {
         return 0;
     }
